+++ src/temperature.jl
@@ -1,4 +1,10 @@
 """
+    offsettemp(::Unit)
+For temperature units, this function is used to set the scale offset.
+"""
+offsettemp(::Unit) = 0
+
+"""
     uconvert{T,D,U}(a::Units, x::Quantity{T,typeof(𝚯),<:TemperatureUnits})
 In this method, we are special-casing temperature conversion to respect scale
 offsets, if they do not appear in combination with other dimensions.